requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
PyYAML==6.0.1
yt-dlp==2023.12.30
